Move files to new folder components
