+++ App.tsx
@@ -1,6 +1,6 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import StopWatch from './src/StopWatch';
+import StopWatch from './src/components/StopWatch';
 
 export default function App() {
   return (
